Removed old Checkbox, old Chip, CheckboxGroup
