--- src/components/Checkbox/Checkbox.module.scss
@@ -1,82 +1,0 @@
-@import '../../../src/theme/old-tokens/shared';
-@import '../../../src/theme/old-tokens/pallete';
-@import '../../../src/theme/old-tokens/functions';
-@import '../../../src/theme/old-tokens/typography';
-
-.checkboxLabel {
-  & > :nth-child(1) {
-    display: block;
-    margin-bottom: 3px;
-    @include text-15-b;
-    &:first-letter {
-      text-transform: capitalize;
-    }
-  }
-}
-
-.checkbox {
-  span.checkboxLabelName {
-    @include text-14;
-    padding-left: 26px;
-    position: relative;
-    line-height: 20px;
-    display: inline-block;
-    margin-bottom: 5px;
-
-    a {
-      color: $green;
-      @include text-14-b;
-    }
-  }
-
-  input + span:before {
-    content: '';
-    width: 16px;
-    height: 16px;
-    display: block;
-    border: 1px solid $back_dark_grey;
-    border-radius: $border-radius;
-    background-color: $white;
-    position: absolute;
-    top: 0;
-    left: 0;
-  }
-
-  input:checked + span:before {
-    border: 1px solid $green;
-    background-color: $green;
-  }
-
-  input:checked + span:after {
-    content: '';
-    width: 8px;
-    height: 4px;
-    display: block;
-    border-left: 2px solid white;
-    border-bottom: 2px solid white;
-    position: absolute;
-    top: 5px;
-    left: 4px;
-    transform: rotate(-45deg);
-  }
-
-  input {
-    display: none;
-  }
-}
-
-.checkboxDisabled {
-  span:before {
-    background-color: $back_grey !important;
-    border: 1px solid $back_dark_grey !important;
-  }
-
-  span:after {
-    border-left: 2px solid $back_dark_grey !important;
-    border-bottom: 2px solid $back_dark_grey !important;
-  }
-}
-
-.required {
-  @include required;
-}
