--- src/components/Checkbox/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Checkbox'
-export * from './MuiCheckbox'
