--- src/components/CheckboxGroup/CheckboxGroup.module.scss
@@ -1,32 +1,0 @@
-@import '../../../src/theme/old-tokens/shared';
-@import '../../../src/theme/old-tokens/pallete';
-@import '../../../src/theme/old-tokens/functions';
-@import '../../../src/theme/old-tokens/typography';
-
-.label {
-  display: flex;
-  & > :nth-child(1) {
-    display: block;
-    margin-bottom: 5px;
-    @include text-15-b;
-    &:first-letter {
-      text-transform: capitalize;
-    }
-  }
-}
-
-.required {
-  @include required;
-}
-
-.errorDescription {
-  @include text-12;
-  color: $error;
-  margin-top: 5px;
-}
-
-.description {
-  margin-top: 2px;
-  color: $color_grey;
-  @include text-12-b;
-}
