--- src/components/Chip/Chip.module.scss
@@ -1,47 +1,0 @@
-@import "../../../src/theme/old-tokens/shared";
-@import "../../../src/theme/old-tokens/pallete";
-@import "../../../src/theme/old-tokens/functions";
-@import "../../../src/theme/old-tokens/typography";
-
-.chip {
-  @include text-12;
-  width: -moz-max-content;
-  width: max-content;
-  padding: 5px 10px;
-  border-radius: 10 * $border_radius;
-}
-
-.grey {
-  background-color: $back_grey;
-  color: $grey_700;
-}
-
-.green {
-  background-color: $light_green;
-  color: $dark_green;
-}
-
-.orange {
-  background-color: $light_orange;
-  color: $orange_500;
-}
-
-.red {
-  background-color: $red_100;
-  color: $red_500;
-}
-
-.blue {
-  background-color: $info_100;
-  color: $info_500;
-}
-
-.text {
-  @include text-10;
-}
-
-.prefix {
-  padding-right: 6px;
-  color: $light_grey;
-  opacity: 0.3;
-}
