--- src/components/NewChip/NewChip.spec.tsx
@@ -1,13 +1,0 @@
-import { render, screen } from '@testing-library/react'
-import React from 'react'
-
-import { NewChip } from './NewChip'
-
-describe('Chip', () => {
-  it('should render the component properly', () => {
-    render(<NewChip label={'text'} addonStart={'prefix'} />)
-
-    expect(screen.getByText('text')).toBeVisible()
-    expect(screen.getByText('prefix')).toBeVisible()
-  })
-})
