+++ src/components/index.ts
@@ -2,9 +2,8 @@ export * from './Alert'
 export * from './Autocomplete'
 export * from './Banner/Banner'
 export * from './Button/Button'
-export { default as Checkbox, MuiCheckbox } from './Checkbox'
-export { default as CheckboxGroup } from './CheckboxGroup'
-export { default as Chip } from './Chip'
+export * from './Checkbox/Checkbox'
+export * from './Chip/Chip'
 export * from './ContactCard/ContactCard'
 export * from './DataItem/DataItem'
 export * from './DataItemList/DataItemsList'
@@ -12,7 +11,6 @@ export * from './Form'
 export { default as ItemCard } from './ItemCard'
 export * from './Layout'
 export { Link } from './Link'
-export * from './NewChip/NewChip'
 export * from './NewTable/NewTable'
 export { default as Pagination } from './Pagination'
 export * from './Progress'
