+++ types/components/Checkbox/Checkbox.d.ts
@@ -1,23 +1,4 @@
+import { CheckboxProps as MuiCheckboxProps } from '@mui/material';
 import React from 'react';
-export type CheckboxOptionType = {
-    name: string;
-    value: string;
-    children?: React.ReactNode;
-};
-type CheckboxOptionTypeOptionalValue = Omit<CheckboxOptionType, 'value'> & {
-    value?: string;
-};
-type CheckboxProps = {
-    name?: string;
-    label?: string;
-    labelClass?: string;
-    option?: CheckboxOptionTypeOptionalValue;
-    description?: string;
-    defaultValue?: boolean;
-    value?: boolean;
-    disabled?: boolean;
-    onClick?: React.ChangeEventHandler<HTMLInputElement>;
-    required?: boolean;
-};
-declare const Checkbox: React.FC<CheckboxProps>;
-export default Checkbox;
+export type CheckboxProps = MuiCheckboxProps;
+export declare const Checkbox: ({ size, sx, ...props }: CheckboxProps) => React.JSX.Element;
