+++ types/components/Chip/Chip.stories.d.ts
@@ -1,25 +1,32 @@
 import React from 'react';
-declare const _default: {
-    title: string;
-    component: React.FC<{
-        text: string;
-        color?: import("./Chip").ChipColorEnum | undefined;
-        prefix?: string | undefined;
-    }>;
-};
+declare const _default: import("@storybook/types").ComponentAnnotations<import("@storybook/react/dist/types-0a347bb9").R, import("@storybook/types").Args>;
 export default _default;
 export declare const Base: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, {
-    text: string;
-    color?: import("./Chip").ChipColorEnum | undefined;
-    prefix?: string | undefined;
-} & {
-    children?: React.ReactNode;
-}>;
+    label: string;
+    size?: ("default" | "small") | undefined;
+    addonStart?: React.ReactNode;
+    color?: "primary" | "error" | "info" | "warning" | "grey" | undefined;
+    onClick?: (() => void) | undefined;
+    isLoading?: boolean | undefined;
+    className?: string | undefined;
+} & React.RefAttributes<HTMLDivElement>>;
+export declare const Sizes: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, any>;
 export declare const Color: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, any>;
-export declare const Prefix: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, {
-    text: string;
-    color?: import("./Chip").ChipColorEnum | undefined;
-    prefix?: string | undefined;
-} & {
-    children?: React.ReactNode;
-}>;
+export declare const AddonStart: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, {
+    label: string;
+    size?: ("default" | "small") | undefined;
+    addonStart?: React.ReactNode;
+    color?: "primary" | "error" | "info" | "warning" | "grey" | undefined;
+    onClick?: (() => void) | undefined;
+    isLoading?: boolean | undefined;
+    className?: string | undefined;
+} & React.RefAttributes<HTMLDivElement>>;
+export declare const isLoading: import("@storybook/types").AnnotatedStoryFn<import("@storybook/react/dist/types-0a347bb9").R, {
+    label: string;
+    size?: ("default" | "small") | undefined;
+    addonStart?: React.ReactNode;
+    color?: "primary" | "error" | "info" | "warning" | "grey" | undefined;
+    onClick?: (() => void) | undefined;
+    isLoading?: boolean | undefined;
+    className?: string | undefined;
+} & React.RefAttributes<HTMLDivElement>>;
